app/requirements.txt
slack_sdk
python-dotenv
slack_sdk
slack_bolt
openai
